moved reading code back into one file
