+++ lib/drivers/medtronicDriver.js
@@ -24,6 +24,9 @@ var TZOUtil = require('../TimezoneOffsetUtil');
 var proc = require('../medtronic/processData');
 var common = require('../commonFunctions');
 
+var crcCalculator = require('../crc.js');
+crcCalculator.crc8_init(0x9b);
+
 var isBrowser = typeof window !== 'undefined';
 var debug = isBrowser ? require('../bows')('MedtronicDriver') : debug;
 
@@ -58,6 +61,21 @@ module.exports = function (config) {
     OPEN_CONNECTION : [0x10,0x01,0x1E]
   };
 
+  var SEND_MESSAGE = [0x12,0x21,0x05];
+
+  var MESSAGES = {
+   SUSPEND : 0x4D,
+   READ_HISTORY : 0x80
+  };
+
+  var getSerial = function() {
+    return '698426'; // FIXME
+  };
+
+  var medtronicHeader = [0xA7,parseInt(getSerial().substring(0,2),16),
+                             parseInt(getSerial().substring(2,4),16),
+                             parseInt(getSerial().substring(4,6),16)];
+
   var messageBuffer = {
     reset: function(){
       this.bytes = new Uint8Array(0);
@@ -77,6 +95,124 @@ module.exports = function (config) {
 
   var probe = function(cb){
     debug('not probing Medtronic');
+  };
+
+  var _sum_lsb = function(bytes) {
+    var sum = 0;
+    bytes.forEach(function (byte) {
+      sum += byte;
+    });
+    return sum & 0xff;
+  };
+
+  var buildMedtronicPacket = function (type, command, parameter) {
+    // first construct payload before we can determine packet length
+    var payload = [];
+    if(command != null) {
+
+      if(parameter != null) {
+        payload = medtronicHeader.concat(command,parameter);
+        var padding = _.fill(new Array(20),0);
+        payload = payload.concat(padding);
+      } else {
+        payload = medtronicHeader.concat(command,0x00);
+        var payloadChecksum = crcCalculator.crc8_checksum(payload);
+        payload = payload.concat(payloadChecksum);
+      }
+    }
+
+    var datalen = 30 + type.length + payload.length;
+    var buf = new ArrayBuffer(datalen + 4); // include 4-byte header
+    var bytes = new Uint8Array(buf);
+
+    var ctr = struct.pack(bytes, 0, '6b6z10b', 0x00, 0x00, 0x00, datalen, 0x51, 0x01, getSerial(),
+                                                0, 0, 0, 0, 0, 0, 0, 0, 0, 0);
+    ctr += struct.copyBytes(bytes, ctr, type, type.length);
+
+    var secondPacketLength = 44; //FIXME
+    if(parameter != null ) {
+      if (command == 0x4D) {
+        ctr += struct.pack(bytes, ctr, '7bi', 0, 0, 0, 0, 0, 0, 0, payload.length+secondPacketLength);
+      } else{
+        //FIXME
+        ctr += struct.pack(bytes, ctr, '7bi', 0,0,0, 0x04,0x10,0x10,0x00,payload.length+secondPacketLength);
+      }
+    } else {
+      ctr += struct.pack(bytes, ctr, '7bi', 0, 0, 0, 0, 0, 0, 0, payload.length);
+    }
+
+    var checkbytes = new Uint8Array(buf.slice(4)); // checksum excludes 4-byte header
+    var ctr2 = struct.copyBytes(checkbytes, ctr - 4, payload, payload.length);
+
+    if(parameter != null) {
+      var secondPacket = buildPaddingPacket(command,parameter).checksum;
+      struct.pack(checkbytes, ctr2 + payload.length + 4, 'b', secondPacket);
+    }
+    var checksum = _sum_lsb(checkbytes);
+
+    ctr += struct.pack(bytes, ctr, 'b', checksum);
+    ctr += struct.copyBytes(bytes, ctr, payload, payload.length);
+
+    console.log('Sending bytes:', common.bytes2hex(bytes));
+    return buf;
+  };
+
+
+  var buildPaddingPacket = function (command, parameter) {
+    var length = 43; //FIXME
+    var padding = _.fill(new Array(length), 0);
+
+    var prevPacketPadding = _.fill(new Array(20), 0);
+    var checkbuf = medtronicHeader.concat(command,parameter,prevPacketPadding,padding);
+    var checksum = crcCalculator.crc8_checksum(checkbuf);
+
+    var datalen = length + 1; // include checksum
+    var buf = new ArrayBuffer(datalen + 4 ); // include 4-byte header
+    var bytes = new Uint8Array(buf);
+
+    var ctr = struct.pack(bytes, 0, '4b', 0x00, 0x00, 0x00, datalen);
+    ctr += struct.copyBytes(bytes, ctr, padding, padding.length);
+    ctr += struct.pack(bytes, ctr, 'b', checksum);
+
+    console.log('Padding packet:', common.bytes2hex(bytes));
+
+    return {command : buf, checksum: checksum};
+  };
+
+  var readModel = function () {
+
+    var cmd = 0x8D;
+
+    return {
+      command: buildMedtronicPacket(SEND_MESSAGE,cmd),
+      parser: function (packet) {
+        var medtronicMessage = packet.slice(33);
+        var messageLength = medtronicMessage[0];
+        var model = struct.extractString(medtronicMessage,1,messageLength);
+        return {model: model};
+      }
+    };
+  };
+
+  var sendCommand = function (cmd) {
+
+    return {
+      command: buildMedtronicPacket(SEND_MESSAGE,cmd),
+      parser: function (packet) {
+        return true;
+      }
+    };
+  };
+
+  var readPage = function (cmd, page) {
+    return {
+      command1: buildMedtronicPacket(SEND_MESSAGE,cmd,page),
+      command2: buildPaddingPacket(cmd,page),
+      parser: function (packet) {
+        var medtronicMessage = packet.slice(33);
+        return {message: medtronicMessage};
+       }
+    };
   };
 
   var extractPacketIntoMessage = function (bytes) {
@@ -190,6 +326,7 @@ module.exports = function (config) {
     hidDevice.send(packet.command1, function () {
       hidDevice.send(packet.command2.command, function () {
 
+        var page  = [];
         var count = 0;
         async.whilst(
             function () { return count < howMany; },
@@ -200,16 +337,18 @@ module.exports = function (config) {
                     return callback(err, null);
                   }
                   var decoded = decodeMessage(packet.parser, result.bytes);
+                  console.log('Part',count,'of page:',decoded);
                   messageBuffer.reset();
+                  page = page.concat(decoded.message);
                   callback(null, count);
                 });
             },
-            function (err, n) {
+            function (err, count) {
               if(err) {
                 return cb(err,null);
               }
-              console.log('Read', n, 'packets of page');
-              return cb(null, true);
+              console.log('Read',count,'parts per page.');
+              return cb(null, page);
             }
         );
       });
@@ -256,7 +395,7 @@ module.exports = function (config) {
           message = extractPacketIntoMessage(packet.slice(MAGIC_HEADER.length));
 
           var header = packet.slice(4,12);
-          if(struct.extractByte(header,0) == 0x51 && _.isEqual(struct.extractString(header,2,6), proc.getSerial())) {
+          if(struct.extractByte(header,0) == 0x51 && _.isEqual(struct.extractString(header,2,6), getSerial())) {
             var payloadLength = struct.extractInt(packet,32);
             console.log('Payload length:', payloadLength);
           }
@@ -322,7 +461,7 @@ module.exports = function (config) {
 
   var openConnection = function () {
     return {
-      command: proc.buildMedtronicPacket(COMMANDS.OPEN_CONNECTION),
+      command: buildMedtronicPacket(COMMANDS.OPEN_CONNECTION),
       parser: function (packet) {
         return null;
       }
@@ -432,7 +571,7 @@ module.exports = function (config) {
           });
         },
         model : function(callback){
-          getOneRecord(proc.readModel(), false, function(err, result) {
+          getOneRecord(readModel(), false, function(err, result) {
             if(err) {
               return cb(err,null);
             }
@@ -459,12 +598,12 @@ module.exports = function (config) {
       progress(0);
       async.series({
         setSuspend : function(callback){
-            getOneRecord(proc.sendCommand(proc.MESSAGES.SUSPEND), true, function (err, result) {
+            getOneRecord(sendCommand(MESSAGES.SUSPEND), true, function (err, result) {
               if(err) {
                 return callback(err,null);
               }
               if(result) {
-                getRecords(proc.readPage(proc.MESSAGES.SUSPEND,[0x01,0x01]), 1, function (err, result) {
+                getRecords(readPage(MESSAGES.SUSPEND,[0x01,0x01]), 1, function (err, result) {
                   if(err) {
                     return callback(err,null);
                   }
@@ -485,17 +624,18 @@ module.exports = function (config) {
           async.whilst(
               function () { return count < 9; },
               function (callback) {
-                  getOneRecord(proc.sendCommand(proc.MESSAGES.READ_HISTORY), true, function (err, result) {
+                  getOneRecord(sendCommand(MESSAGES.READ_HISTORY), true, function (err, result) {
                     if(err) {
                       return callback(err,null);
                     }
                     if(result) {
-                      getRecords(proc.readPage(proc.MESSAGES.READ_HISTORY,[0x01,count]), 4, function (err, result) {
+                      getRecords(readPage(MESSAGES.READ_HISTORY,[0x01,count]), 4, function (err, result) {
                         if(err) {
                           return callback(err,null);
                         }
                         if(result) {
                           count++;
+                          console.log("PAGE:", result);
                           // TODO: do something with result
                           return callback(null, count);
                         } else {
